CELOSTAR-397: Improve performance for weekday calendar (#1555)
## Why I'm doing:

 We have observed slow weekday calendars for Poste Italiane Spa.

## What I'm doing:

This is a simple improvement to the `compute_overlap` function for the
weekday calendar case.

The idea behind this improvement is that no loop is required to compute
the overlap for dates that are far apart. We can compute the overlap in
the first week, the overlap in the last week and multiple the overlap of
all weeks in between. Visually it looks like this:

![image](https://github.com/user-attachments/assets/afa4ad4a-4574-4a34-8ac8-345577b9865e)
- In black you can see the timeline split into weeks.
- In red you see the TimeRange that is repeated for each week
- In orange you see the left and right timestamp for which we want to
compute the overlap
- In green you see the overlap that needs to be added.
- For the first week (week -2) we compute the overlap between `left` and
the end of the week.
- For the last week (week 2) we compute the overlap between the
beginning of the week and `right` (which is 0 in the example)
- For the weeks in between (weeks -1, 0, 1) we multiply the the size of
the time range by 3


Benchmark shows that this about 3.5X faster.

**Note for reviewer: Only the `compute_overlap` function changed, the
rest is just moving the code around.**

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5