Add saola .clang-format to src/exprs/celonis/modules (#1678)
## Why I'm doing:

Celostar and Saola use different clang-format rules. This means applying
the celostar clang-format to code copied from saola causes clang-format
to reformat the entire file.

## What I'm doing:

As far as I can tell most of the copied saola code is located in
`be/exprs/celonis/modules`. As a stop gap to stop clang-format from
completely reformatting these files, but to also enable some formatting
for these files copy the saola .clang-format file there so it is used by
the formatter for any files in that directory and below.

See the minimal changes to `align_model.cpp` to see that the saola style
is picked up correctly.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5