[Enhancement] Change CELONIS_ADD_TIMEUNITS_CALENDAR to handle constant time_unit in prepare (#1706)
## Why I'm doing:

Currently CELONIS_ADD_TIMEUNITS_CALENDAR only handles constant Calendar
column. Constant time_unit column is not handled in prepare.

## What I'm doing:

This PR changes CELONIS_ADD_TIMEUNITS_CALENDAR to also handle constant
time_unit column in prepare method. It also update add_timeunits_bench
and add_timeunits_calendar_bench.

Based on the benchmark of add_timeunits (i.e., calendar is empty), we
got

| Operation | Average Improvement |
|-----------|-------------------|
| AddMilliseconds | 33.1% |
| AddSeconds | 36.7% |
| AddMinutes | 41.7% |
| AddHours | 47.5% |
| AddDays | 62.3% |
| AddWorkdays | 55.8% |
| **Overall Average** | **45.1%** |

For add_timeunits with Calendar, the previous benchmark is no a
different hardware, it is hard to compare. And the previous benchmark
result is outdated for weekday calendar (it is too slow). We redo the
benchmark on the same hardware and got
https://gist.github.com/zhangyaqi1989/454ff33281040e1ef378c719d0014394.
For factory calendar, we saw 17.54% improvement.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5