[BugFix] Put all weekdays in the same calendar (#1733)
## Why I'm doing:

`weekday_calendar` creates a separate entry in the calendar for each day
with the same calendar_id.
This causes an issue when validating the calendar during
`make_intersect_calendar`, as there it's expected each calendar_id
occurs once in the weekday calendar.

## What I'm doing:

Making sure that all the weekdays are put into the same calendar entry
in `weekday_calendar`.
Adjusting tests to validate this is working. 

This issue was uncovered by `intersectWeekdayFactoryCalendarTest`
compliance test, see
[here](https://docs.google.com/spreadsheets/d/1NQDUZSWIdLySWIEGMzqtSNa03pdlnjcX-GzIV3W66AM/edit?gid=1766774618#gid=1766774618).



Fixes #issue

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5