saola consistency: remove object counts from MO_BPMN_GRAPH and IM result (#1739)
## Why I'm doing:

We observed test differences between Saola and CS-SR due to mismatching
object counts returned by `MO_BPMN_GRAPH` (see [CS pull
12283](https://github.com/celonis/celostar/pull/12283)). The root cause
is that we "removed" the object counts (they are set to 0 instead of
actually removing the column from the result structure) some time ago in
Saola ([Saola pull
10527](https://github.com/celonis/cpm-query-engine/pull/10527)).

## What I'm doing:

This PR removes the object count computation in the IM computation in
the CPML proxy, as these are the counts used to populate the
`MO_BPMN_GRAPH` result tables. I checked and if I see correctly these
counts returned by IM should not be used anywhere else.

## What type of PR is this:

- [x] BugFix (not really a bug but an inconsistency between CS-SR and
Saola)
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [x] Miscellaneous: upgrade & downgrade compatibility, etc.

Values in the `OBJECT_COUNT` column of the `bpmn_edges` table as
returned by `MO_BPMN_GRAPH` are set to 0 with this change.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5