[Enhancement] Change CELONIS_MAKE_WORKDAY_CALENDAR to order the calendar entries before output (deterministic output and reduce ~25% output size) (#1763)
## Why I'm doing:

Currently the output of CELONIS_MAKE_FACTORY_CALENDAR is not
deterministic (the order of the calendar entries depends on the
execution). This is not ideal.

## What I'm doing:

This PR changes CELONIS_MAKE_WORKDAY_CALENDAR to order the calendar
entries before output. This makes its output deterministic. Also sorting
the proto entries before serialization results in better compression. In
the test case (6000 rows in workday_table), the size of output is
reduced from about 42300 bytes to 31092 bytes.

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5