[BugFix] Fix compilation issue of calendar benchmarks (#1784)
## Why I'm doing:

Recently we change the signature of `to_base64_encoded_string` to
support string compression. This makes the benchmark of some calendar
functions can not compile.

## What I'm doing:

This PR fixes the compilation issue and enables string compression in
the benchmarks. One interesting observation is the
remap_timestamps_calendar_prepare_bench is ~1.5x slower, this is because
in the benchmark the calendar is generated randomly (no pattern, high
entropy), compression does not help. We could change the custom calendar
functions to not compress the calendar when the compression is not
effective (e.g., input is small, input has high entropy).

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5