[BugFix] Add ARRAY_INT to calc_throughput and reorder functions in the python script (#66)
- [x] BugFix
- [x] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

The way SR picks the best function for the set of parameters is by
iterating over the list of functions with the name
'celonis_calc_throughput' and picking the first one that works. Since
INT can be cast to VARCHAR, having ARRAY_VARCHAR-version in the list of
functions first leads to SR casting the ARRAY_INT input to
ARRAY_VARCHAR. Therefore, I changed the order of the functions in the
python script.
Also note that the other array functions are ordered from "smaller"
types (INT or bool) to "larger" types (VARCHAR and JSON), i think for
the same reason.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function