Revert null_to_empty function (#178)
- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [X] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

<!-- (Please describe the changes you have made. In which scenarios will
this bug be triggered and what measures have you taken to fix the bug?)
-->

As I understand, the current behavior of null_to_empty function is same
to getting data_column from a nullable column and the whole function can
be replaced with 1 line
FunctionHelper::get_data_column_of_nullable(columns[0]) which also has
just 4 lines of code.

null_to_empty function generates new offsets to handle null as an empty
array by appending the same offset of the previous entry if an array is
null. But actually the original offsets are also made like this and the
original and the result offsets are same.

I've attached the screenshot showing the input column of null_to_empty
unittest. The input offsets already have a repeating "3".

![Screenshot-ArrayNull](https://user-images.githubusercontent.com/117620956/233150983-7b1085df-b6d2-42de-a046-debcb207ce0d.png)

Unless there are other values to provide with the function, we may
delete the function. Or am I missing something?
What do you think?

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be auto
backported to target branch
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3
  - [ ] 2.2