[Feature] Support non-constant match in celonis_in() (#191)
- [ ] BugFix
- [X] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

<!--
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
_If PR is about `failing-tests or flakes`, please post the related
issues/tests in a comment and do not use `Fixes`_*
-->
Fixes #

<!-- (Please describe the changes you have made. In which scenarios will
this bug be triggered and what measures have you taken to fix the bug?)
-->

Added support for non-constant match to celonis_in().
- While the function supports both constant and non-constant matches,
the function is called always with non-constant matches at the moment in
local build test. It looks like starrocks does not pass constant array
column somehow. I've asked about it in starrocks channel.
- InReproTest in celostar was passed.

Moved TYPE_VARCHAR to the last in functions.py so that other types are
not cast to TYPE_VARCHAR.

- [X] I have added test cases for my bug fix or my new feature
- [ ] This pr will affect users' behaviors
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function

- [ ] I have checked the version labels which the pr will be auto
backported to target branch
  - [ ] 2.5
  - [ ] 2.4
  - [ ] 2.3
  - [ ] 2.2