[BugFix] ALIGN_MODEL alignment activity join incorrect if trace contains nulls PMT-1581 (#1677)
## Why I'm doing:

If a trace contains nulls the previous strategy of just ignoring these
activities would cause us to compute the join between the ACTIVITY and
ALIGNMENT table incorrectly. E.g. for a trace <A,null,B> ignoring nulls
would yield the variant <A,B>. In the algorithm if we would want to join
a move to the activity B we just used the variant index (in this case 1)
and incorrectly join to the null activity.

## What I'm doing:

Fixes the issue described above by introducing a mapping from indices
into the variant to an index in the case.
See https://github.com/celonis/cpm-query-engine/pull/11760 for the
equivalent fix in Saola.

Also fixed some clang-tidy issues on the side.
Also formatted some files that needed minor adjustments by clang-format.
Unfortunately saola and celostar use different formatting, so the
formatter also wants to reformat the files copied from the saola
implementation. Did not reformat those. Maybe `align_model_helper.cpp`
should be reformatted. Did not do that to keep the diff cleaner

## What type of PR is this:

- [X] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5