[BugFix] Register match_strings_prepare and match_strings_close in BE (CELONIS_MATCH_STRINGS) (#1306)
This bug was introduced in
[PR](https://github.com/celonis/celostar-starrocks/pull/1292) which adds
prepare() method for CELONIS_MATCH_STRINGS.

Without registering the `prepare` and `close` methods, SR framework will
not call prepare and the execution of match_strings will fail.

Why this is not caught by unit tests and benchmark?
In unit tests and benchmark, prepare is called manually.

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5