Refactor celonis_array_sources_targets operator - CELOSTAR-482 (#1622)
- Preparation for https://celonis.atlassian.net/browse/CELOSTAR-431
- Remove code duplication
- For example, we create a separate function
celonis_array_sources_targets, which is called by celonis_array_sources
and celonis_array_targets. This not only removes code duplication, but
also enables us to assign the celonis_array_sources_target function
(with a template parameter for the edge configuration) to
state->function in the prepare function.

In preparation for supporting all edge configurations in
celonis_array_sources/target, we should clean up the code of this
operator a bit, as it contains a lot of duplicate code fragments.

- [ ] BugFix
- [ ] Feature
- [ ] Enhancement
- [x] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5