fix release workflow: be images added to fe manifest (#1886)
## Why I'm doing:
When attempting to pull a `FE` container from ghcr (i.e.
`ghcr.io/celonis/celostar/starrocks-fe:gh-branch-3.3-2f0f22dd-2025-08-28T07-05-38`)
I kept getting a container which looked like a `BE`. It had all of the
`be_entrypoint` scripts and was missing the `fe_entrypoint`.

Having a closer look at the release workflow I suspect we are amending
the (multi-arch) FE manifest with BE containers instead of FE. This was
only the case for ghcr and not for ECR, so to my understanding this
should have had no prod impact.

**Note that I am not familiar with this workflow, so not sure if this is
the root cause for the issues I am seeing. I just thought this bit
looked suspicious, so please review carefully :)**

## What I'm doing:

Amend the ghcr multi-arch FE manifest with the correct FE containers in
the release workflow.

Fixes #issue

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [x] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5

Co-authored-by: ji.hu <ji.hu@celonis.com>