[BugFix] Change CELONIS_LIKE to not use regex match for pattern without wildcard (#1898)
## Why I'm doing:

In the original implementation, there is a `std::shared_ptr<re2::RE2>
pattern_re2;` in the FragmentLocalState. This can cause contention in
production (slow query, majority of the time spend on a mutex:
RWLocker). We also found that RE2::PartialMatch is an overkill for
substring match.

## What I'm doing:

This PR changes CELONIS_LIKE to not use regex match for pattern without
wildcard. We found that it is 2x faster than using RE2::PartialMatch
when the string is short (len = 20). It has similar performance as
RE2::PartialMatch when the string is long (len = 60).

## What type of PR is this:

- [x] BugFix
- [ ] Feature
- [ ] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [x] Yes, this PR will result in a change in behavior.
- [ ] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression
evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with
different default values
- [ ] Policy changes: use new policy to replace old one, functionality
automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [ ] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or
behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [ ] I have checked the version labels which the pr will be
auto-backported to the target branch
  - [ ] 3.3
  - [ ] 3.2
  - [ ] 3.1
  - [ ] 3.0
  - [ ] 2.5